Create backup of a database
This script uses the mysqldump command to create a backup of the mydatabase database, and saves it to a file named mydatabase_yyyymmdd.sql. Then it uses the scp command to copy the backup file to a remote location. The script then removes the local copy of the backup file.

You can schedule the script to run daily, weekly or monthly using cron, so that you always have a recent backup of your databases. you can use the command crontab -e to open the cronjob editor and add a schedule for the script to run.